Web_DB_project/requirements.txt
Flask
psycopg
flask-cors
python-dotenv
Flask-SQLAlchemy

